Using ransack for resources search and sorting.
This also uses Alchemy::Resource and resources controller for all internal rsources Alchemy has.
