+++ lib/alchemy/engine.rb
@@ -12,6 +12,7 @@ require 'jquery-rails'
 require 'jquery-ui-rails'
 require 'kaminari'
 require 'non-stupid-digest-assets'
+require 'ransack'
 require 'request_store'
 require 'responders'
 require 'sass-rails'
